refactor(web): consolidate project creation components
- Refactor `StepBasics` to use
  `StepBasicsGeneralSettings` and
  `StepBasicsSourceSection` for better
  organization.
- Remove redundant preset and build settings
  components.
- Add `slugifyProjectName` helper.
- Update `CreateProjectInput` type definitions.
