+++ apps/web/src/utils/slugify.ts
@@ -1,0 +1,7 @@
+export function slugifyProjectName(value: string) {
+	return value
+		.toLowerCase()
+		.replace(/[^a-z0-9-]/g, '-')
+		.replace(/-+/g, '-')
+		.replace(/^-|-$/g, '');
+}
